# [GPU] Convolution: improve small IC/OC performance