--- a/src/gpu/intel/jit/conv/lookup_table_data.cpp
+++ b/src/gpu/intel/jit/conv/lookup_table_data.cpp
@@ -1,5 +1,5 @@
 /*******************************************************************************
-* Copyright 2023-2024 Intel Corporation
+* Copyright 2023-2025 Intel Corporation
 *
 * Licensed under the Apache License, Version 2.0 (the "License");
 * you may not use this file except in compliance with the License.
@@ -31,13 +31,9 @@ const char** get_conv_lookup_table_entries() {
         "xehpg dpas bwd_w x16:any:x16 mb1 ic3ih224oc128oh56kh4sh4ph0 simd=8 l=oh8 T=oc2 i=ic4kw4oc8ow16 bufs=0",
         "xehpg dpas fwd x8:x8:any mb128b ic64iw3136oc256ow3136kw1pw0 simd=8 l=ic2 T=ow8 i=ic32mb32oc32 bufs=2",
         "xehpg dpas fwd x16:x16:any mb2 ic320iw2304oc640ow2304kw1pw0 simd=8 l=ic20 T=oc4ow8 i=ic16oc32ow24",
-        "xehpg dpas fwd x16:x16:any mb32b ic1ih80iw100650oc1oh80ow100650kh1kw23ph0pw11 simd=8 l=kw3 T=ow8 i=ic2kw8mb32oc8 bufs=1",
-        "xehpg dpas fwd x8:x8:any mb32b ic1ih80iw100650oc1oh80ow100650kh1kw23ph0pw11 simd=8 l=kw3 T=ow8 i=ic4kw8mb32oc8 bufs=1",
         "xehpg dpas fwd x16:x16:any mb2 ic1280ih24oc1280oh24kh3ph1 simd=8 l=ic80kh3kw3 T=oc4 i=ic16oc32ow24",
         "xehpg dpas fwd x16:x16:any mb2 ic960iw9216oc320ow9216kw1pw0 simd=8 l=ic60 T=oc4ow8 i=ic16oc32ow32",
         "xehpg dpas fwd x8:x8:any mb128b ic128ih56oc128oh28kh3sh2ph1 simd=8 l=ic4kh3kw3 T=oc4ow8 i=ic32mb32oc32 bufs=3",
-        "xehpg mad fwd x16:x16:any mb1 ic1ih80iw100650oc1oh80ow100650kh1kw23ph0pw11 simd=8 l=kw3 T=ow8 i=kw8oc8ow16 bufs=1",
-        "xehpg mad fwd x8:x8:any mb1 ic1ih80iw100650oc1oh80ow100650kh1kw23ph0pw11 simd=8 l=x T=ow2 i=kw23oc8ow15 bufs=0",
         "xehpg dpas fwd x8:x8:any mb128b ic64ih56oc64oh56kh3ph1 simd=8 l=ic2kh3kw3 T=oc2ow8 i=ic32mb32oc32 bufs=3",
         "xehpg dpas fwd x8:x8:any mb128b ic1024iw196oc256ow196kw1pw0 simd=8 l=ic32 T=oc8ow4 i=ic32mb32oc32 bufs=3",
         "xehpc dpas fwd x16:x16:any mb1 ic256ih5oc12oh5kh3ph1 simd=16 l=ic2kh3kw3 T=ic8 i=ic16oc16ow8 bufs=0",
@@ -306,8 +302,6 @@ const char** get_conv_lookup_table_entries() {
         "xehpc dpas fwd x8:x8:any mb32b ic3id8ih224iw224oc16od8oh112ow112kd1kh3kw3sh2sw2pd0ph1pw1 simd=16 l=kh3 T=x i=ic4kw8mb32oc16 bufs=3",
         "xehpg dpas fwd x16:x16:any mb1 ic640ih32oc640oh16kh3sh2ph1 simd=8 l=ic10kh3kw3 T=ic4 i=ic16oc32ow16 bufs=0",
         "xehpg dpas fwd x16:x16:any mb1 ic1280ih8oc1280oh8kh3ph1 simd=8 l=ic20kh3kw3 T=ic2 i=ic32oc16ow8 bufs=0",
-        "xehpc dpas fwd x8:x8:any mb32b ic1ih80iw100650oc1oh80ow100650kh1kw23ph0pw11 simd=16 l=kw3 T=ow4 i=ic4kw8mb32oc16 bufs=3",
-        "xehpc dpas fwd x16:x16:any mb32b ic1ih80iw100650oc1oh80ow100650kh1kw23ph0pw11 simd=16 l=kw3 T=ow8 i=ic2kw8mb32oc16 bufs=3",
         "xehpc dpas bwd_w x16:any:x16 mb64 ic256iw90oc256ow59kw32pw0 simd=16 l=mb64ow4 T=ic8oc4 i=ic16oc64ow16",
         "xehpg dpas bwd_w x16:any:x16 mb3 ic256ih10iw16oc256oh10ow16kh5kw5ph2pw2 simd=8 l=oh10 T=ic8oc4 i=ic16oc64ow16 bufs=3",
         "xehpg dpas fwd x16:x16:any mb1 ic1920iw256oc1280ow256kw1pw0 simd=8 l=ic120 T=oc8ow4 i=ic16oc16ow32",
@@ -335,8 +329,6 @@ const char** get_conv_lookup_table_entries() {
         "xehpc mad fwd f64:f64:any mb2048 ic2ih10oc8oh8kh3ph0 simd=8 l=kh3 T=x i=ic2kw3oc8ow8 bufs=0",
         "xehpc dpas bwd_w x16:any:x16 mb64 ic128iw16oc64ow15kw2pw0 simd=16 l=x T=ic8 i=ic8kw2oc16ow16",
         "xehpg dpas fwd x8:x8:any mb32b ic10ih15oc20oh13kh3ph0 simd=8 l=kh3kw2 T=oc4ow4 i=ic16kw2mb32oc8 bufs=3",
-        "xehpc mad fwd x8:x8:any mb1 ic1ih80iw100650oc1oh80ow100650kh1kw23ph0pw11 simd=16 l=x T=ow4 i=kw23oc16ow16 bufs=0",
-        "xehpc mad fwd x16:x16:any mb1 ic1ih80iw100650oc1oh80ow100650kh1kw23ph0pw11 simd=16 l=x T=x i=kw23oc16ow16 bufs=0",
         "xehpg dpas fwd x16:x16:any mb1 ic1280ih10oc1280oh8kh3ph0 simd=8 l=ic20kh3kw3 T=ic2 i=ic32oc16ow8 bufs=0",
         "xehpc dpas bwd_d any:x16:x16 mb64 ic256iw90oc256ow59kw32pw0 simd=16 l=kw32oc8 T=ic8iw4 i=ic32iw30oc32",
         "xehpg dpas fwd x16:x16:any mb2 ic640ih24oc1280oh24kh3ph1 simd=8 l=ic40kh3kw3 T=oc4 i=ic16oc32ow24",