Jira: [MFDNN-13199](https://jira.devtools.intel.com/browse/MFDNN-13199)

List of changes:

- Adjusted heuristics to enable `mad` (instead of `dpas`) and M/N transposition
    - Both optimizations are to reduce padding (and related redundant work) in case of small channels
    - This is like 10th or so version of the heuristics - unfortunately still seeing a few regressions in P3 KPIs. There are much more improvements so it does not look too bad
- Removed suboptimal lookup table entries
- Refactored and simplified accumulator type and transposed blocking logic
    - We still incorrectly use f16 accumulator with f16 data type in some cases. I tried to fix it but it causes quite a few model-level slowdowns so let's keep it this way for now. In the future we need to add FP math mode based dispatching between f16 vs f32 (easier part) and ask our users to port their code to use the proper FP math mode to avoid regressions (harder part)

Jira cases:

| Case                                                                                                                    | base perf (GOps/sec) | test perf (GOps/sec) | Ratio |
|-------------------------------------------------------------------------------------------------------------------------|----------------------|----------------------|-------|
| --dir=FWD_I --dt=f16:f16:f16 --stag=acdb --dtag=acdb mb1ic3ih1080iw1920oc1oh1080ow1920kh1kw1ph0pw0                      | 10.935               | 86.114               | 7.88  |
| --dir=FWD_I --dt=f16:f16:f16 --stag=abcd --dtag=acdb --attr-post-ops=relu mb1ic3ih1080iw1920oc3oh1080ow1920kh1kw1ph0pw0 | 47.573               | 103.411              | 2.17  |