@echeresh 2025-02-14
make test perf-gpu
set primitive=conv deconv