@echeresh 2025-02-19
make test perf-gpu
set primitive=conv deconv