@echeresh 2025-02-26
make test perf-gpu
set primitive=conv deconv