@echeresh 2025-02-27
make test perf-gpu
set primitive=conv deconv