@echeresh 2025-03-04
make test perf-gpu
set primitive=conv deconv